Document ProtoCBOR specification
JSON is a web-interchange format that is widely used but has inefficient
space usage properties. The CBOR format (RFC8949) is a newer web
standard for exchanging JSON-like objects with better encoding
overheads.

Whereas protocol buffers are still proprietary to Google,
CBOR is an open standard. For platforms that provide multiple encodings
of their messages through protoc for JSON and Protobuf, it is
challenging to introduce a CBOR format without disruption unless CBOR is
a supported target for the protocol buffer compiler.

TODO: replace with link
This proposal will come with a prototype implementation for Golang to
prove the concept and engage the protocol buffer community to discuss
whether full support across target languages is something they're
willing to accept maintenance for, at least temporarily as a beta
experiment for early adopters to try.
The goal is to not merge this specification without first accepting the
initial implementation. The initial implementation should not be merged
until the flagship project that uses it (in-toto framework) demonstrates
its correct behavior.